# designs/contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: "教育培训机构教务收费管理系统 API（EduFeeMS）"
  description: >
    教育培训机构教务收费管理系统（EduFeeMS）RESTful API 接口契约文档。
    涵盖认证、学员管理、课程管理、班级管理、收费管理、考勤管理六大业务域。
  version: "1.0.0"
  contact:
    name: "EduFeeMS 开发团队"
servers:
  - url: "http://localhost:8080/api"
    description: "本地开发环境"
  - url: "https://edu-fee-ms.example.com/api"
    description: "生产环境"
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: "用户登录"
      description: "使用用户名和密码登录，返回JWT Token"
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              username: "admin"
              password: "Admin@123456"
      responses:
        "200":
          description: "登录成功，返回Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: "用户名或密码错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      tags:
        - auth
      summary: "获取当前用户信息"
      description: "根据JWT Token获取当前登录用户的详细信息"
      responses:
        "200":
          description: "成功获取用户信息"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoResponse"
        "401":
          description: "未登录或Token已过期"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== 学员管理 ====================
...
  /students/{id}:
    get:
      tags:
        - students
      summary: "获取学员详情"
      description: "根据ID获取学员详细信息，含班级信息和缴费记录摘要"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: "学员ID"
      responses:
        "200":
          description: "查询成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetailResponse"
        "404":
          description: "学员不存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - students
      summary: "修改学员信息"
      description: "更新学员基本信息（不含分班/转班操作）"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: "学员ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdateRequest"
            example:
              name: "张三"
              phone: "13800138001"
              address: "北京市海淀区XX路XX号"
              remark: "已完成基础阶段"
      responses:
        "200":
          description: "修改成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentResponse"
...
  /students/{id}/transfer:
    post:
      tags:
        - students
      summary: "学员转班"
      description: "将学员从当前班级转到目标班级"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: "学员ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
            example:
              targetClassId: 2
              transferDate: "2026-07-15"
              reason: "课程时间冲突"
      responses:
        "200":
          description: "转班成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
  /students/{id}/withdraw:
    post:
      tags:
        - students
      summary: "学员退学"
      description: "标记学员退学，需关联退费流程"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: "学员ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequest"
            example:
              withdrawDate: "2026-07-20"
              reason: "家庭搬迁"
              requireRefund: true
      responses:
        "200":
          description: "退学申请成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawResponse"

  # ==================== 课程管理 ====================
...
  /fees/payment:
    post:
      tags:
        - fees
      summary: "学费收取"
      description: "为学员创建收费记录，支持全款/分期收取"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCreateRequest"
            example:
              studentId: 1
              classId: 1
              amount: 4800.00
              paymentMethod: "WECHAT_PAY"
              installmentPlan: "FULL"
              remark: "全款缴纳"
      responses:
        "200":
          description: "收费成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          description: "金额异常或学员状态不符"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /fees/records:
    get:
      tags:
        - fees
      summary: "收费台账查询"
      description: "查询收费台账明细，支持多维度筛选"
      parameters:
        - name: pageNum
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: "学员ID"
        - name: classId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: "班级ID"
        - name: paymentStatus
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PaymentStatus"
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          description: "流水号/学员姓名模糊搜索"
      responses:
        "200":
          description: "查询成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecordPageResponse"
  /fees/dunning:
    get:
      tags:
        - fees
      summary: "欠费查询"
      description: "查询欠费学员列表，支持按欠费金额区间筛选"
      parameters:
        - name: pageNum
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: classId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: minAmount
          in: query
          required: false
          schema:
            type: number
          description: "最低欠费金额"
        - name: maxAmount
          in: query
          required: false
          schema:
            type: number
          description: "最高欠费金额"
        - name: overdueDays
          in: query
          required: false
          schema:
            type: integer
          description: "逾期天数下限"
      responses:
        "200":
          description: "查询成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DunningPageResponse"

  # ==================== 考勤管理 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT认证，登录后获取Token，请求时携带 Authorization: Bearer <token>"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: "错误码"
          example: 401
        message:
          type: string
          description: "错误信息"
          example: "用户名或密码错误"
        data:
          type: object
          nullable: true
          example: null
...
    PageInfo:
      type: object
      properties:
        pageNum:
          type: integer
          description: "当前页码"
        pageSize:
          type: integer
          description: "每页条数"
        total:
          type: integer
          format: int64
          description: "总记录数"
        pages:
          type: integer
          description: "总页数"

    # ---- 认证相关 ----
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: "用户名"
          minLength: 2
          maxLength: 50
          example: "admin"
        password:
          type: string
          description: "密码"
          format: password
          minLength: 6
          maxLength: 128
          example: "Admin@123456"
    LoginResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "登录成功"
        data:
          type: object
          properties:
            token:
              type: string
              description: "JWT Token"
              example: "eyJhbGciOiJIUzI1NiIs..."
            tokenType:
              type: string
              example: "Bearer"
            expiresIn:
              type: integer
              description: "过期时间（秒）"
              example: 7200
            userInfo:
              $ref: "#/components/schemas/UserInfo"
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: "admin"
        realName:
          type: string
          example: "系统管理员"
        role:
          type: string
          enum: ["ADMIN", "FINANCE", "ACADEMIC", "TEACHER"]
          example: "ADMIN"
        phone:
          type: string
          example: "138****8000"
    UserInfoResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "操作成功"
        data:
          $ref: "#/components/schemas/UserInfo"

    # ---- 学员相关 ----
    StudentStatus:
      type: string
      enum:
        - ENROLLED
        - IN_CLASS
        - WITHDRAWN
        - GRADUATED
      description: "学员状态：ENROLLED-已报名, IN_CLASS-在读, WITHDRAWN-已退学, GRADUATED-已毕业"
...
    StudentUpdateRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        gender:
          type: string
          enum: ["MALE", "FEMALE"]
        birthday:
          type: string
          format: date
        parentName:
          type: string
        parentPhone:
          type: string
        address:
          type: string
        remark:
          type: string
    StudentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        gender:
          type: string
        phone:
          type: string
          description: "手机号（列表场景脱敏）"
        birthday:
          type: string
          format: date
        parentName:
          type: string
        parentPhone:
          type: string
        address:
          type: string
        source:
          type: string
        status:
          $ref: "#/components/schemas/StudentStatus"
        createTime:
          type: string
          format: date-time
          example: "2026-06-22 10:30:00"
    StudentResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "操作成功"
        data:
          $ref: "#/components/schemas/StudentDTO"
    StudentDetailResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "操作成功"
        data:
          allOf:
            - $ref: "#/components/schemas/StudentDTO"
            - type: object
              properties:
                currentClass:
                  $ref: "#/components/schemas/ClassDTO"
                enrollmentHistory:
                  type: array
                  items:
                    $ref: "#/components/schemas/EnrollmentRecordDTO"
                recentPayments:
                  type: array
                  items:
                    $ref: "#/components/schemas/FeeRecordDTO"
...
    EnrollmentRecordDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        classId:
          type: integer
          format: int64
        className:
          type: string
        enrollDate:
          type: string
          format: date
        status:
          type: string
          enum: ["ACTIVE", "TRANSFERRED", "WITHDRAWN"]
...
    TransferRequest:
      type: object
      required:
        - targetClassId
      properties:
        targetClassId:
          type: integer
          format: int64
        transferDate:
          type: string
          format: date
        reason:
          type: string
          maxLength: 200
    TransferResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "转班成功"
        data:
          type: object
          properties:
            transferId:
              type: integer
              format: int64
            fromClassId:
              type: integer
              format: int64
            toClassId:
              type: integer
              format: int64
    WithdrawRequest:
      type: object
      required:
        - withdrawDate
        - reason
      properties:
        withdrawDate:
          type: string
          format: date
        reason:
          type: string
          maxLength: 200
        requireRefund:
          type: boolean
          default: true
          description: "是否需要退费"
    WithdrawResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "退学申请已提交"
        data:
          type: object
          properties:
            withdrawId:
              type: integer
              format: int64
            refundRequired:
              type: boolean
            estimatedRefundAmount:
              type: number
              description: "预估退费金额"

    # ---- 课程相关 ----
...
    ClassStatus:
      type: string
      enum: ["RECRUITING", "IN_PROGRESS", "COMPLETED"]
      description: "班级状态：RECRUITING-招生中, IN_PROGRESS-开课中, COMPLETED-已结课"
...
    ClassDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        courseId:
          type: integer
          format: int64
        courseName:
          type: string
        capacity:
          type: integer
        enrolledCount:
          type: integer
          description: "已报名人数"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        teacherName:
          type: string
        classroom:
          type: string
        status:
          $ref: "#/components/schemas/ClassStatus"
        createTime:
          type: string
          format: date-time
...
    PaymentStatus:
      type: string
      enum: ["PAID", "PARTIAL", "UNPAID", "REFUNDED"]
      description: "缴费状态：PAID-已缴清, PARTIAL-部分缴纳, UNPAID-未缴, REFUNDED-已退费"
    PaymentCreateRequest:
      type: object
      required:
        - studentId
        - classId
        - amount
        - paymentMethod
      properties:
        studentId:
          type: integer
          format: int64
        classId:
          type: integer
          format: int64
        amount:
          type: number
          description: "收费金额（元）"
          minimum: 0.01
        paymentMethod:
          type: string
          enum: ["CASH", "WECHAT_PAY", "ALIPAY", "BANK_TRANSFER", "POS"]
        installmentPlan:
          type: string
          enum: ["FULL", "INSTALLMENT"]
          default: "FULL"
          description: "缴费方式：FULL-全款, INSTALLMENT-分期"
        installments:
          type: integer
          description: "分期数（installmentPlan=INSTALLMENT时必填）"
        remark:
          type: string
          maxLength: 200
    PaymentResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "收费成功"
        data:
          type: object
          properties:
            paymentId:
              type: integer
              format: int64
            serialNo:
              type: string
              description: "业务流水号"
              example: "FEE-20260622-000001"
            amount:
              type: number
            paymentTime:
              type: string
              format: date-time
...
    FeeRecordDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        serialNo:
          type: string
          description: "业务流水号"
          example: "FEE-20260622-000001"
        studentId:
          type: integer
          format: int64
        studentName:
          type: string
        classId:
          type: integer
          format: int64
        className:
          type: string
        amount:
          type: number
          description: "金额（元）"
        paymentMethod:
          type: string
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        paymentTime:
          type: string
          format: date-time
        operatorName:
          type: string
          description: "操作人"
    FeeRecordPageResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "操作成功"
        data:
          type: object
          properties:
            pageInfo:
              $ref: "#/components/schemas/PageInfo"
            summary:
              type: object
              properties:
                totalAmount:
                  type: number
                  description: "总金额合计"
                paidAmount:
                  type: number
                  description: "已收金额合计"
                unpaidAmount:
                  type: number
                  description: "未收金额合计"
            records:
              type: array
              items:
                $ref: "#/components/schemas/FeeRecordDTO"
    DunningItemDTO:
      type: object
      properties:
        studentId:
          type: integer
          format: int64
        studentName:
          type: string
        studentPhone:
          type: string
        className:
          type: string
        courseName:
          type: string
        totalFee:
          type: number
        paidAmount:
          type: number
        unpaidAmount:
          type: number
          description: "欠费金额"
        overdueDays:
          type: integer
          description: "逾期天数"
    DunningPageResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "操作成功"
        data:
          type: object
          properties:
            pageInfo:
              $ref: "#/components/schemas/PageInfo"
            totalUnpaidAmount:
              type: number
              description: "欠费总额"
            records:
              type: array
              items:
                $ref: "#/components/schemas/DunningItemDTO"

    # ---- 考勤相关 ----
...